Add new bfs
