+++ srcs/count_turn.c
@@ -6,7 +6,7 @@
 /*   By: thle <thle@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 11:15:03 by thle              #+#    #+#             */
-/*   Updated: 2022/11/01 15:02:46 by itkimura         ###   ########.fr       */
+/*   Updated: 2022/11/02 17:02:44 by itkimura         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,8 +97,7 @@ t_bool		count_turn(t_info *info, t_path *list, int count_path, int *curr_turn)
 		if (*curr_turn < path[i][0] - 1 + path[i][1])
 			*curr_turn = path[i][0] - 1 + path[i][1];
 	}
-	for (int i = 0; i < count_path ; i++)
-		printf("path[0] = %d path[1] = %d total:%d\n", path[i][0], path[i][1], path[i][0] + path[i][1]);
-	printf("curr_turn = %d\n", *curr_turn);
+	//for (int i = 0; i < count_path ; i++)
+	//	printf("path[0] = %d path[1] = %d total:%d\n", path[i][0], path[i][1], path[i][0] + path[i][1]);
 	return (TRUE);
 }
